+++ src/frontend/apps/impress/src/features/docs/doc-header/components/AlertNetwork.tsx
@@ -75,7 +75,12 @@ export const AlertNetworkModal = ({ onClose }: AlertNetworkModalProps) => {
       onClose={() => onClose()}
       rightActions={
         <>
-          <Button aria-label={t('OK')} onClick={onClose} color="error">
+          <Button
+            aria-label={t('OK')}
+            onClick={onClose}
+            color="error"
+            autoFocus
+          >
             {t('I understand')}
           </Button>
         </>