+++ src/frontend/apps/impress/src/features/docs/doc-versioning/components/ModalSelectVersion.tsx
@@ -132,6 +132,7 @@ export const ModalSelectVersion = ({
                 </Text>
                 <ButtonCloseModal
                   aria-label={t('Close the version history modal')}
+                  autoFocus
                   onClick={onClose}
                   size="nano"
                 />